Extract functions for logging to system log/stderr
It makes it easier to understand the function `usdbmuxd_log` because now
it distinguishs only once between the two log types. The code for each
type of log is now linear and therefore easier to read. The only common
code of the two log types is the allocation and deallocation of memory
for the extended message format. I copied these two lines for the sake
of simplicity. This makes it also possible to adjust the malloc to the
different length of the log format.